@@ -12,7 +12,7 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": "^8.0",
+        "php": "^8.1",
         "symfony/framework-bundle": "^6.2",
         "symfony/twig-bundle": "^6.2",
         "claviska/simpleimage": "^3.3",
