@@ -3,6 +3,7 @@
 namespace Trexima\EuropeanCvBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
+use Doctrine\Persistence\Proxy;
 use Symfony\Component\Validator\Constraints as Assert;
 use Trexima\EuropeanCvBundle\Entity\Embeddable\YearRange;
 use Trexima\EuropeanCvBundle\Entity\Enum\EducationTypeEnum;
@@ -55,19 +56,16 @@ class EuropeanCVEducation
         return $this->id;
     }
 
-    public function setId(?int $id): void
-    {
-        $this->id = $id;
-    }
-
-    public function getEuropeanCV(): EuropeanCV
+    public function getEuropeanCV(): ?EuropeanCV
     {
         return $this->europeanCV;
     }
 
-    public function setEuropeanCV(EuropeanCV $europeanCV): void
+    public function setEuropeanCV(?EuropeanCV $europeanCV): self
     {
         $this->europeanCV = $europeanCV;
+
+        return $this;
     }
 
     public function getType(): ?EducationTypeEnum
@@ -75,9 +73,11 @@ class EuropeanCVEducation
         return $this->type;
     }
 
-    public function setType(?EducationTypeEnum $type): void
+    public function setType(?EducationTypeEnum $type): self
     {
         $this->type = $type;
+
+        return $this;
     }
 
     public function getTitle(): ?string
@@ -85,9 +85,11 @@ class EuropeanCVEducation
         return $this->title;
     }
 
-    public function setTitle(?string $title): void
+    public function setTitle(?string $title): self
     {
         $this->title = $title;
+
+        return $this;
     }
 
     public function getKovCode(): ?string
@@ -95,9 +97,11 @@ class EuropeanCVEducation
         return $this->kovCode;
     }
 
-    public function setKovCode(?string $kovCode): void
+    public function setKovCode(?string $kovCode): self
     {
         $this->kovCode = $kovCode;
+
+        return $this;
     }
 
     public function getKovTitle(): ?string
@@ -105,9 +109,11 @@ class EuropeanCVEducation
         return $this->kovTitle;
     }
 
-    public function setKovTitle(?string $kovTitle): void
+    public function setKovTitle(?string $kovTitle): self
     {
         $this->kovTitle = $kovTitle;
+
+        return $this;
     }
 
     public function getYearRange(): YearRange
@@ -115,9 +121,11 @@ class EuropeanCVEducation
         return $this->yearRange;
     }
 
-    public function setYearRange(YearRange $yearRange): void
+    public function setYearRange(YearRange $yearRange): self
     {
         $this->yearRange = $yearRange;
+
+        return $this;
     }
 
     public function getDescription(): ?string
@@ -125,9 +133,11 @@ class EuropeanCVEducation
         return $this->description;
     }
 
-    public function setDescription(?string $description): void
+    public function setDescription(?string $description): self
     {
         $this->description = $description;
+
+        return $this;
     }
 
     public function __clone()
@@ -137,7 +147,14 @@ class EuropeanCVEducation
          * https://www.doctrine-project.org/projects/doctrine-orm/en/2.6/cookbook/implementing-wakeup-or-clone.html
          */
         if ($this->id) {
-            $this->setId(null);
+            if ($this instanceof Proxy && !$this->__isInitialized()) {
+                $this->__load();
+            }
+
+            $this->id = null;
+
+            $yearRange = clone $this->getYearRange();
+            $this->setYearRange($yearRange);
         }
     }
 }
