@@ -3,6 +3,7 @@
 namespace Trexima\EuropeanCvBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
+use Doctrine\Persistence\Proxy;
 use Symfony\Component\Validator\Constraints as Assert;
 use Trexima\EuropeanCvBundle\Entity\Enum\LanguageEnum;
 use Trexima\EuropeanCvBundle\Entity\Enum\LanguageLevelEnum;
@@ -11,7 +12,7 @@ use Trexima\EuropeanCvBundle\Entity\Enum\LanguageLevelEnum;
  * EuropeanCV language
  */
 #[ORM\Table(name: 'european_cv_language')]
-#[ORM\Index(name: 'language_idx', columns: ['language'])]
+#[ORM\Index(columns: ['language'], name: 'language_idx')]
 #[ORM\Entity]
 class EuropeanCVLanguage
 {
@@ -30,7 +31,13 @@ class EuropeanCVLanguage
             new Assert\NotNull(message: 'trexima_european_cv.level_not_empty')
         ]
     )]
-    #[ORM\Column(type: 'string', nullable: true, length: 2, enumType: LanguageEnum::class, options: ['comment' => 'ISO 639-1'])]
+    #[ORM\Column(
+        type: 'string',
+        length: 2,
+        nullable: true,
+        enumType: LanguageEnum::class,
+        options: ['comment' => 'ISO 639-1'],
+    )]
     private ?LanguageEnum $language = null;
 
     #[Assert\When(
@@ -47,19 +54,16 @@ class EuropeanCVLanguage
         return $this->id;
     }
 
-    public function setId(?int $id): void
-    {
-        $this->id = $id;
-    }
-
     public function getEuropeanCV(): ?EuropeanCV
     {
         return $this->europeanCV;
     }
 
-    public function setEuropeanCV(EuropeanCV $europeanCV): void
+    public function setEuropeanCV(?EuropeanCV $europeanCV): self
     {
         $this->europeanCV = $europeanCV;
+
+        return $this;
     }
 
     public function getLanguage(): ?LanguageEnum
@@ -67,9 +71,11 @@ class EuropeanCVLanguage
         return $this->language;
     }
 
-    public function setLanguage(?LanguageEnum $language): void
+    public function setLanguage(?LanguageEnum $language): self
     {
         $this->language = $language;
+
+        return $this;
     }
 
     public function getLevel(): ?LanguageLevelEnum
@@ -77,9 +83,11 @@ class EuropeanCVLanguage
         return $this->level;
     }
 
-    public function setLevel(?LanguageLevelEnum $level): void
+    public function setLevel(?LanguageLevelEnum $level): self
     {
         $this->level = $level;
+
+        return $this;
     }
 
     public function __clone()
@@ -89,7 +97,11 @@ class EuropeanCVLanguage
          * https://www.doctrine-project.org/projects/doctrine-orm/en/2.6/cookbook/implementing-wakeup-or-clone.html
          */
         if ($this->id) {
-            $this->setId(null);
+            if ($this instanceof Proxy && !$this->__isInitialized()) {
+                $this->__load();
+            }
+
+            $this->id = null;
         }
     }
 }
